Added a basic quiz setup using js
Using js in combination with an html file, I set up a quiz that tests users on past tense versions of words. I also renamed unit1exercise to unitexercise, and I will be making it the exercise page for all units, but changing the page based on which unit is chosen.