Made a number of changes to make the exercise setup better, described below.
Added significantly more choices of words for the exercise, made the exercise randomly choose 10 words to quiz the user on with each visit of the page/reset with reset button. Also added scoring for the exercise and checks/x's to help the user know if they got a question right or wrong. Also added some css for these checks/x's and for the new reset button.