Connect Exercises to Database #40 : Updated code to allow dynamic user access to exercise points
Currently having issues with accessing the database, I am not sure what is going wrong exactly, it works for my fetch but not when I try to modify the userAccount. However, I added capabilty to increment the points for all users depending on the login, and I added a function to fetch the current points to display on the exercises page.